Hero Select Screen Adapted to Designs
The Hero Select now reflects our current design. Hero Descriptions,
Passives and abilities (text) are loaded in from a txt file. The
Descriptions must be in this order 1. Eir 2. Heimdal 3. Skadi 4. Thor 5.
Sam.  If they are not in this order, they will not match the correct
character. They must be separated by a semi-colon (the last description
should not have a semi-colon).