Improved Victory Feedback and Ability Fix
If a player achieves an elimination victory, the game now ends as soon
as they kill the last unit, not on the next turn. Additionally, it now
displays the correct player for this victory.

Persistent AoE passives are also now properly removed from units under
their effects if the hero dies.