GUI, Capture, and Particle Updates
Shifted temporary GUI text onto actual GUI so it stays on camera when it
zooms. Added some more feedback for when altars are captured. Changed
the capture function so it will only run on button command. Added some
simple, and awful particles to the shrines so they stand out a bit more.