Island, Particles, and Scene Tweaks
Brought in and textured the new island, and added the remaining
particles. Adjusted the GUI so some text is no longer obscured. resized
all the heroes and units. Particles themselves still need to be properly
adjusted to be smaller and less obscuring. Added a skybox to the scene
also. Still need an ocean.